refactor: split large inline test modules into sibling files
Move the inline #[cfg(test)] mod tests out of the three biggest files this work
touched into sibling tests.rs modules (foo.rs + foo/tests.rs), cutting the logic
file readers open:
- checkpoint_coordinator.rs 4641 -> 2550 (tests -> checkpoint_coordinator/tests.rs)
- delta.rs 2555 -> 1645 (tests -> delta/tests.rs)
- delta_io.rs 2146 -> 1347 (tests -> delta_io/tests.rs)

Pure relocation (use super::* still resolves); behavior unchanged. db.rs was
already split; pipeline_lifecycle.rs has no test module.
